AR Tap to place changes
I want to create a spawn of object when the + sign is added. Attempting to make one object to be the default active, then change to smth else when the rest are tapped. This is so that people won't have to do 2 interaction when the camera is running and they want to place the object it will automatically come out?